update car api to get a car type while fetching cars; add state to TabBar to evict recall the api
