+++ flutter_cars/lib/app/pages/home/home_page.dart
@@ -1,102 +1,47 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_cars/app/pages/home/cars_listview.dart';
 import 'package:flutter_cars/app/pages/home/drawer_list.dart';
 import 'package:flutter_cars/data/services/car_api.dart';
-import 'package:flutter_cars/data/services/models/Car.dart';
 
-class HomePage extends StatelessWidget {
+class HomePage extends StatefulWidget{
+  @override
+  _HomePageState createState() => _HomePageState();
+}
+
+class _HomePageState extends State<HomePage> with SingleTickerProviderStateMixin<HomePage> {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Carros'),
-      ),
-      body: _body(),
-      drawer: DrawerList(),
-    );
-  }
-
-  _body() {
-    final Future<List<Car>> carsFuture = CarApi.fetchCars();
-
-    return FutureBuilder(
-      future: carsFuture,
-      builder: (context, snapshot) {
-        if (snapshot.hasError) {
-          return Center(
-            child: Text(
-              "An error occurred while fetching data from server",
-              style: TextStyle(
-                color: Colors.red,
-                fontSize: 22,
+    return DefaultTabController(
+      length: 3,
+      child: Scaffold(
+        appBar: AppBar(
+          title: Text('Carros'),
+          bottom: TabBar(
+            tabs: <Widget>[
+              Tab(
+                text: "Classics",
               ),
-            ),
-          );
-        }
-        if (!snapshot.hasData) {
-          return Center(
-            child: CircularProgressIndicator(),
-          );
-        }
-        final List<Car> cars = snapshot.data;
-        return _listView(cars);
-      },
-    );
-  }
-
-  Container _listView(List<Car> cars) {
-    return Container(
-      margin: EdgeInsets.all(16),
-      child: ListView.builder(
-        itemCount: cars != null ? cars.length : 0,
-        itemBuilder: (context, index) {
-          final Car car = cars[index];
-          return Card(
-            color: Colors.grey[100],
-            child: Container(
-              padding: EdgeInsets.all(10),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: <Widget>[
-                  Center(
-                    child: car.urlPhoto != null
-                        ? Image.network(
-                            car.urlPhoto,
-                            width: 250,
-                          )
-                        : Image.network(
-                            "https://cdn0.iconfinder.com/data/icons/shift-travel/32/Speed_Wheel-512.png",
-                            width: 150,
-                          ),
-                  ),
-                  Text(
-                    car.name ?? "Pistons",
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                    style: TextStyle(fontSize: 25),
-                  ),
-                  Text(
-                    "description...",
-                    style: TextStyle(fontSize: 14),
-                  ),
-                  ButtonTheme.bar(
-                    child: ButtonBar(
-                      children: <Widget>[
-                        FlatButton(
-                          child: const Text('DETAILS'),
-                          onPressed: () {},
-                        ),
-                        FlatButton(
-                          child: const Text('SHARE'),
-                          onPressed: () {},
-                        )
-                      ],
-                    ),
-                  ),
-                ],
+              Tab(
+                text: "Sportives",
               ),
+              Tab(
+                text: "Lux",
+              ),
+            ],
+          ),
+        ),
+        body: TabBarView(
+          children: <Widget>[
+            CarsListView(
+              carType: CarType.classic,
             ),
-          );
-        },
+            CarsListView(carType: CarType.sport),
+            CarsListView(
+              carType: CarType.lux,
+            ),
+          ],
+        ),
+        drawer: DrawerList(),
       ),
     );
   }
